requirements.txt
streamlit
plotly
pandas
openpyxl
numpy
supabase
python-dotenv